Update PBGuidelines2012.tex
Updated section 1.1 numbered arg table, and included paragraph about function tags. If we need more information than this paragraph, it would be more appropriate to create a new section for it later in the document. As it is, section 1.1 is the most relevant place for it, but I have concerns about putting too much detail in there. Also, updated section 1.2.1 'leave' examples.